semana-02/dia-05/requirements.txt
flask
firebase_admin
python-dotenv
pyrebase4
requests-toolbelt==0.10.1